feat: Update: 1.Kafka components and add ut.

+++ kafka-plus-core/src/main/java/io/github/photowey/kafka/plus/core/clients/builder/admin/topic/NewTopicBuilder.java
@@ -33,7 +33,7 @@ public interface NewTopicBuilder {
 
     NewTopicBuilder numPartitions(Integer numPartitions);
 
-    NewTopicBuilder replicationFactor(Short replicationFactor);
+    NewTopicBuilder replicationFactor(Integer replicationFactor);
 
     NewTopicBuilder replicasAssignments(Map<Integer, List<Integer>> replicasAssignments);
